Fixed a few warnings.

+++ src/global.h
@@ -203,7 +203,7 @@ void *alloca();
 #endif
 
 #define MAX_INT32 2147483647
-#define MIN_INT32 -2147483648
+#define MIN_INT32 -(INT32)2147483648
 
 #define INT16 short
 #define INT8 char
